backend/requirements/requirements.txt
django==4.1.5
django-environ==0.9.0
psycopg2==2.9.5
djangorestframework==3.14.0
drf-spectacular==0.25.1  # https://github.com/tfranzel/drf-spectacular